+++ src/app/app.routes.ts
@@ -8,6 +8,11 @@ export const routes: Routes = [
     path: '',
     children: [
       {
+        path: '',
+        redirectTo: 'dashboard',
+        pathMatch: 'full',
+      },
+      {
         path: 'dashboard',
         component: DashboardComponent,
       },