+++ src/app/shared/models/tag.ts
@@ -1,0 +1,6 @@
+export interface Tag {
+  id: string;
+  label: string;
+  tooltip: string;
+  disabled: boolean;
+}