+++ src/app/store/movies.selector.ts
@@ -2,6 +2,7 @@ import { createSelector } from '@ngrx/store';
 import { MoviesState } from './movies.reducer';
 import { Movie } from '../shared/models/movie';
 import { Favorite } from '../shared/models/favorites';
+import { Tag } from '../shared/models/tag';
 
 export interface AppState {
   movies: MoviesState;
@@ -22,4 +23,9 @@ export const selectMoviesData = createSelector(
 export const selectFavorites = createSelector(
   selectMovies,
   (state: MoviesState): Favorite[] => state.favorites
+);
+
+export const selectTags = createSelector(
+  selectMovies,
+  (state: MoviesState): Tag[] => state.tags
 );