+++ changelog.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.4] - 2024-01-07
 
+### Change
+
+- Default route
+
 ### Added
 
 - No data component